/api/users/{u}/collections を別controllerに移動
